unify xy processing with country processing
- XYGeofabrik class takes raw into in _init_ method
- get_xy_coordinates_from_input therefore moved to geofabrik.py
- unittests adopted to new way of calling XYGeofabrik class
